fix(ci): python-full-coverage — split auth_mounts_claude pin/lock helpers into their own module
``tests/unit/test_core_decomposition_maintainability.py::
test_first_party_code_files_stay_under_line_limit`` failed because
``src/awf/node/auth_mounts_claude.py`` grew to 1518 lines, past the 1500-line
first-party guardrail, as the overlay-probe fixes landed on this branch.

Move the coherent base-signature-pin + provision-lock unit — the
``CLAUDE_AUTH_OVERLAY_BASE_PIN_WRITE_FAILED`` /
``CLAUDE_AUTH_OVERLAY_PROVISION_LOCK_UNAVAILABLE`` reason codes, the
``.overlay.lock`` name, the would-block errnos, ``_pinned_overlay_base``,
``_pin_matches_signature``, ``_record_overlay_base_pin``,
``_live_overlay_pin_signature`` and ``_overlay_provision_lock`` — into a new
``awf.node.auth_mounts_claude_pin`` module, following the existing
``auth_mounts_claude_*`` decomposition. ``auth_mounts_claude`` re-imports every
name, so ``awf.node.auth_mounts.<name>`` stays the stable import surface and
behavior is unchanged.

``fcntl`` now lives in the new module, so the two ``flock`` monkeypatches in the
overlay lock tests patch it there instead of on ``auth_mounts_claude``.
